requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.0.1
ansi2html==1.7.0
arch==5.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.1
catboost==1.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dash==2.5.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.3
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
graphviz==0.20.1
idna==3.3
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==4.11.4
inflection==0.5.1
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-dash==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lightgbm==4.0.0
lxml==4.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mlxtend==0.22.0
more-itertools==9.1.0
multidict==6.0.4
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.9.2
nest-asyncio==1.5.5
nltk==3.2.4
notebook==6.4.11
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.1.1
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pigar==2.1.1
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.8.2
preprocessing==0.1.13
prometheus-client==0.14.1
prompt-toolkit==3.0.29
property-cached==1.6.4
protobuf==4.24.2
psutil==5.9.1
pure-eval==0.2.2
pyaml==23.5.9
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-certifi-win32==1.6.1
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
Quandl==3.7.0
requests==2.27.1
retrying==1.3.3
rich==13.5.2
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.8.1
seaborn==0.12.1
Send2Trash==1.8.0
setuptools-scm==7.0.5
six==1.16.0
sketchpy==0.0.15
sklearn==0.0.post1
smmap==5.0.0
soupsieve==2.3.2.post1
sphinx-rtd-theme==0.2.4
stack-data==0.2.0
statsmodels==0.13.5
streamlit==1.26.0
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
tenacity==8.2.3
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.2.2.post1
turtle==0.0.1
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.9
validators==0.21.2
virtualenv==20.14.1
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
xgboost==1.7.6
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.8.0
